Major rewrite of most text.
Added two extra figures ("between diversity" and a sketch of a track
plot) and moved one figure within text.
Rewrite is mostly going for a slightly more minimalist and purely
diversity-centered design.
I think the case is fairly convincing as it is, but there is still room
for using persistence statistics to say something about other kinds of
reassortments.