Adding dated railroad track figure.
Because of uncertainty trait switch timing along a branch I've used the
angle of lineages reassorting into a different background to indicate
uncertainty.
The start of a reassortment indicates the mean estimated date of a
reassortant clade's parent node and end is the reassortant clade's mean
tmrca. HPDs would be better, but also a lot wider.
Angles seem like a good idea for conveying uncertainty in dates if
someone ever decides to develop a program for making railroad track
plots.