Adding cover letter.
A few points on initial submission:

1. eLife asks for a single paragraph cover letter, current one has 4
(because it's mimicking Trevor's flux cover letter). Having said that,
no one at eLife seemed to mind Trevor's cover letter. Should we cut it
down to the first paragraph and let the abstract do the rest or keep it
as is?

2. eLife love love loves early-career scientists and I might be
worrying prematurely here, but I can't see any easy way of specifying
that I'm a PhD student (which is about as early as it gets) other than
putting my title in as PhD student Gytis Dudas the First.

3. eLife also prefers the format where Methods go near the end, rather
than after the Introduction. Is it worth changing the manuscript to
reflect this?

4. The manuscript is also 900 words over eLife's preferred limit of
5000 words. Shall we ignore this or make minimal attempts at throwing
out some embellishments?

If all goes well everything can be submitted on the week beginning on 3
March (the week after next).