Add LD circles, new LineageRatiosOverTime.
LD circles might be a nice looking alternative to the mean LD figure
(which is a bit of a dodgy concept). Also D' figures are a proof of
concept - it's a ridiculous metric of LD for flu.

Also added new lineageRatios figure that shows how many genomes were
used to infer ratios in the 452 dataset and the black lines indicate the
ratio in the 1600 dataset (the numbers  in brackets are numbers of
genomes from the 1600 dataset used to find the small black line ratios).