Fix #40, #39
Clarify deltaTMRCA confusion. I've decided that deltaTMRCA is best
thought of as the mean of absolute deviations in TMRCAs between 2
trees. We get normalized deltaTMRCA by quantifying deltaTMRCAs within
(=phylogenetic noise) and between (=phylogenetic noise + signal) trees.
The result of this is a measure of how similar the signal is to noise,
which in our case is the desired outcome. Hopefully that makes sense.
Also move all the LD figures to supplementary. It was done as a
confirmation of results and also addresses the figure number problem.